+++ firmware/config/boards/hellen/hellen128/board_configuration.cpp
@@ -10,12 +10,8 @@
  * @author Andrey Belomutskiy, (c) 2012-2020
  */
 
-#include "global.h"
-#include "engine.h"
-#include "engine_math.h"
-#include "allsensors.h"
+#include "pch.h"
 #include "fsio_impl.h"
-#include "engine_configuration.h"
 
 static void hellenWbo() {
 	engineConfiguration->enableAemXSeries = true;
@@ -115,6 +111,8 @@ void setBoardConfigOverrides(void) {
 
 	engineConfiguration->canTxPin = GPIOD_1;
 	engineConfiguration->canRxPin = GPIOD_0;
+
+	engineConfiguration->vrThreshold[0].pin = GPIOD_14;
 }
 
 void setPinConfigurationOverrides(void) {