+++ firmware/controllers/algo/engine.cpp
@@ -9,19 +9,15 @@
  * @author Andrey Belomutskiy, (c) 2012-2020
  */
 
-#include "engine.h"
-#include "allsensors.h"
-#include "efi_gpio.h"
-#include "pin_repository.h"
+#include "pch.h"
+
 #include "trigger_central.h"
 #include "fuel_math.h"
-#include "engine_math.h"
 #include "advance_map.h"
 #include "speed_density.h"
 #include "advance_map.h"
 #include "os_util.h"
 #include "os_access.h"
-#include "settings.h"
 #include "aux_valves.h"
 #include "map_averaging.h"
 #include "fsio_impl.h"
@@ -29,20 +25,16 @@
 #include "backup_ram.h"
 #include "idle_thread.h"
 #include "idle_hardware.h"
-#include "sensor.h"
 #include "gppwm.h"
 #include "tachometer.h"
 #include "dynoview.h"
 #include "boost_control.h"
 #include "fan_control.h"
 #include "ac_control.h"
+#include "vr_pwm.h"
 #if EFI_MC33816
  #include "mc33816.h"
 #endif // EFI_MC33816
-
-#if EFI_TUNER_STUDIO
-#include "tunerstudio_outputs.h"
-#endif /* EFI_TUNER_STUDIO */
 
 #if EFI_PROD_CODE
 #include "trigger_emulator_algo.h"
@@ -225,6 +217,8 @@ void Engine::periodicSlowCallback(DECLARE_ENGINE_PARAMETER_SIGNATURE) {
 	updateSlowSensors(PASS_ENGINE_PARAMETER_SIGNATURE);
 	checkShutdown(PASS_ENGINE_PARAMETER_SIGNATURE);
 
+	updateVrPwm(PASS_ENGINE_PARAMETER_SIGNATURE);
+
 #if EFI_FSIO
 	runFsio(PASS_ENGINE_PARAMETER_SIGNATURE);
 #else
@@ -315,7 +309,7 @@ void Engine::updateSwitchInputs(DECLARE_ENGINE_PARAMETER_SIGNATURE) {
 #if EFI_GPIO_HARDWARE
 	// this value is not used yet
 	if (isBrainPinValid(CONFIG(clutchDownPin))) {
-		engine->clutchDownState = efiReadPin(CONFIG(clutchDownPin));
+		engine->clutchDownState = CONFIG(clutchDownPinInverted) ^ efiReadPin(CONFIG(clutchDownPin));
 	}
 	if (hasAcToggle(PASS_ENGINE_PARAMETER_SIGNATURE)) {
 		bool result = getAcToggle(PASS_ENGINE_PARAMETER_SIGNATURE);
@@ -326,7 +320,7 @@ void Engine::updateSwitchInputs(DECLARE_ENGINE_PARAMETER_SIGNATURE) {
 		engine->acSwitchState = result;
 	}
 	if (isBrainPinValid(CONFIG(clutchUpPin))) {
-		engine->clutchUpState = efiReadPin(CONFIG(clutchUpPin));
+		engine->clutchUpState = CONFIG(clutchUpPinInverted) ^ efiReadPin(CONFIG(clutchUpPin));
 	}
 	if (isBrainPinValid(CONFIG(throttlePedalUpPin))) {
 		engine->engineState.idle.throttlePedalUpState = efiReadPin(CONFIG(throttlePedalUpPin));
@@ -365,10 +359,6 @@ void Engine::reset() {
 	 */
 	engineCycle = getEngineCycle(FOUR_STROKE_CRANK_SENSOR);
 	memset(&ignitionPin, 0, sizeof(ignitionPin));
-	for (int camIndex = 0;camIndex < CAMS_PER_BANK;camIndex++) {
-		// todo: is it possible to make it constructor argument?
-		vvtTriggerConfiguration[camIndex].index = camIndex;
-	}
 }
 
 
@@ -633,7 +623,7 @@ float Engine::getTimeIgnitionSeconds(void) const {
 	// return negative if the ignition is turned off
 	if (ignitionOnTimeNt == 0)
 		return -1;
-	float numSeconds = (float)NT2US(getTimeNowNt() - ignitionOnTimeNt) / 1000000.0f;
+	float numSeconds = (float)NT2US(getTimeNowNt() - ignitionOnTimeNt) / US_PER_SECOND_F;
 	return numSeconds;
 }