+++ firmware/controllers/algo/engine.h
@@ -70,15 +70,18 @@ protected:
 
 class VvtTriggerConfiguration final : public TriggerConfiguration {
 public:
-	VvtTriggerConfiguration() : TriggerConfiguration("TRG ") {}
-	// todo: is it possible to make 'index' constructor argument?
-	int index = 0;
+	const int index;
+
+	VvtTriggerConfiguration(const char * prefix, const int index) : TriggerConfiguration(prefix), index(index) {
+	}
 
 protected:
 	bool isUseOnlyRisingEdgeForTrigger() const override;
 	bool isVerboseTriggerSynchDetails() const override;
 	trigger_type_e getType() const override;
 };
+
+#define DEFAULT_MOCK_SPEED -1
 
 class Engine final : public TriggerStateListener {
 public:
@@ -103,8 +106,24 @@ public:
 
 	GearControllerBase *gearController;
 
+
+	float mockVehicleSpeed = DEFAULT_MOCK_SPEED; // in kilometers per hour
+
+	efitick_t vssLastSignalTimeNt = 0;
+	efitick_t vssDiff = 0;
+
+	efitick_t mostRecentSparkEvent;
+	efitick_t mostRecentTimeBetweenSparkEvents;
+	efitick_t mostRecentIgnitionEvent;
+	efitick_t mostRecentTimeBetweenIgnitionEvents;
+
 	PrimaryTriggerConfiguration primaryTriggerConfiguration;
-	VvtTriggerConfiguration vvtTriggerConfiguration[CAMS_PER_BANK];
+#if CAMS_PER_BANK == 1
+	VvtTriggerConfiguration vvtTriggerConfiguration[CAMS_PER_BANK] = {{"VVT1 ", 0}};
+#else
+	VvtTriggerConfiguration vvtTriggerConfiguration[CAMS_PER_BANK] = {{"VVT1 ", 0}, {"VVT2 ", 1}};
+#endif
+
 	efitick_t startStopStateLastPushTime = 0;
 
 #if EFI_SHAFT_POSITION_INPUT