+++ firmware/controllers/algo/obd_error_codes.h
@@ -19,6 +19,7 @@ extern "C"
 
 // Back in the day we wanted enums to be 32 bit integers.
 // as of 2020 preference is with ' __attribute__ ((__packed__))' allowing one-byte enums
+#define ENUM_16_BITS 60000
 #define ENUM_32_BITS 2000000000
 
 typedef enum {
@@ -1982,7 +1983,7 @@ typedef enum {
 	CUSTOM_ERR_6628 = 6628,
 	CUSTOM_STACK_6629 = 6629,
 
-	CUSTOM_IGN_MATH_STATE = 6630,
+	CUSTOM_ERR_6030 = 6630,
 	CUSTOM_ERR_6631 = 6631,
 	CUSTOM_ERR_6632 = 6632,
 	CUSTOM_ANGLE_NAN = 6633,