+++ firmware/controllers/bench_test.cpp
@@ -21,7 +21,7 @@
  */
 
 // todo: rename this file
-#include "global.h"
+#include "pch.h"
 
 #if EFI_ENGINE_CONTROL
 #if !EFI_UNIT_TEST
@@ -29,19 +29,12 @@
 #include "os_access.h"
 #include "flash_main.h"
 #include "bench_test.h"
-#include "io_pins.h"
 #include "main_trigger_callback.h"
-#include "engine_configuration.h"
-#include "pin_repository.h"
-#include "efi_gpio.h"
-#include "settings.h"
 #include "idle_thread.h"
 #include "periodic_thread_controller.h"
-#include "tps.h"
 #include "electronic_throttle.h"
 #include "cj125.h"
 #include "malfunction_central.h"
-#include "tunerstudio_outputs.h"
 #include "trigger_emulator_algo.h"
 #include "microsecond_timer.h"
 
@@ -320,7 +313,7 @@ static void handleBenchCategory(uint16_t index) {
 	case CMD_TS_BENCH_FAN_RELAY:
 		fanBench();
 		return;
-	case CMD_TS_BENCH_AC_FAN_RELAY:
+	case CMD_TS_BENCH_FAN_RELAY_2:
 		fan2Bench();
 		return;
 	default:
@@ -385,6 +378,25 @@ static void handleCommandX14(uint16_t index) {
 #endif
 	case 0x12:
 		widebandUpdatePending = true;
+		return;
+	case 0x14:
+#ifdef STM32F7
+		void sys_dual_bank(void);
+		/**
+		 * yes, this would instantly cause a hard fault as a random sequence of bytes is decoded as instructions
+		 * and that's the intended behavious - the point is to set flash properly and to re-flash once in proper configuration
+		 */
+		sys_dual_bank();
+		rebootNow();
+#else
+		firmwareError(OBD_PCM_Processor_Fault, "Unexpected dbank command", index);
+#endif
+		return;
+	case 0x15:
+#if EFI_PROD_CODE
+		extern bool burnWithoutFlash;
+		burnWithoutFlash = true;
+#endif // EFI_PROD_CODE
 		return;
 	default:
 		firmwareError(OBD_PCM_Processor_Fault, "Unexpected bench x14 %d", index);