+++ firmware/controllers/trigger/decoders/trigger_structure.cpp
@@ -19,9 +19,9 @@
  */
 
 
-#include "global.h"
+#include "pch.h"
+
 #include "os_access.h"
-#include "engine.h"
 #include "trigger_chrysler.h"
 #include "trigger_gm.h"
 #include "trigger_nissan.h"
@@ -40,9 +40,6 @@
 #if EFI_SENSOR_CHART
 #include "sensor_chart.h"
 #endif /* EFI_SENSOR_CHART */
-
-#include "engine_configuration.h"
-		extern persistent_config_container_s persistentState;
 
 void event_trigger_position_s::setAngle(angle_t angle DECLARE_ENGINE_PARAMETER_SUFFIX) {
 	findTriggerPosition(&ENGINE(triggerCentral.triggerShape),
@@ -159,7 +156,10 @@ angle_t TriggerWaveform::getAngle(int index) const {
 	int crankCycle = index / privateTriggerDefinitionSize;
 	int remainder = index % privateTriggerDefinitionSize;
 
-	return getCycleDuration() * crankCycle + getSwitchAngle(remainder);
+	auto cycleStartAngle = getCycleDuration() * crankCycle;
+	auto positionWithinCycle = getSwitchAngle(remainder);
+
+	return cycleStartAngle + positionWithinCycle;
 }
 
 void TriggerWaveform::addEventClamped(angle_t angle, trigger_wheel_e const channelIndex, trigger_value_e const stateParam, float filterLeft, float filterRight) {
@@ -366,7 +366,7 @@ void TriggerWaveform::setTriggerSynchronizationGap3(int gapIndex, float syncRati
 /**
  * this method is only used on initialization
  */
-uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, float target) const {
+uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, angle_t targetAngle) const {
 	size_t engineCycleEventCount = getLength();
 
 	efiAssert(CUSTOM_ERR_ASSERT, engineCycleEventCount != 0 && engineCycleEventCount <= 0xFFFF, "engineCycleEventCount", 0);
@@ -382,9 +382,9 @@ uint16_t TriggerWaveform::findAngleIndex(TriggerFormDetails *details, float targ
         int middle = (left + right) / 2;
 		angle_t eventAngle = details->eventAngles[middle];
 
-        if (eventAngle < target) {
+        if (eventAngle < targetAngle) {
             left = middle + 1;
-        } else if (eventAngle > target) {
+        } else if (eventAngle > targetAngle) {
             right = middle - 1;
         } else {
             // Values are equal
@@ -418,7 +418,10 @@ void findTriggerPosition(TriggerWaveform *triggerShape,
 
 	int triggerEventIndex = details->triggerIndexByAngle[(int)angle];
 	angle_t triggerEventAngle = details->eventAngles[triggerEventIndex];
-	if (angle < triggerEventAngle) {
+	angle_t offsetFromTriggerEvent = angle - triggerEventAngle;
+
+	// Guarantee that we aren't going to try and schedule an event prior to the tooth
+	if (offsetFromTriggerEvent < 0) {
 		warning(CUSTOM_OBD_ANGLE_CONSTRAINT_VIOLATION, "angle constraint violation in findTriggerPosition(): %.2f/%.2f", angle, triggerEventAngle);
 		return;
 	}
@@ -428,7 +431,7 @@ void findTriggerPosition(TriggerWaveform *triggerShape,
 		chibios_rt::CriticalSectionLocker csl;
 
 		position->triggerEventIndex = triggerEventIndex;
-		position->angleOffsetFromTriggerEvent = angle - triggerEventAngle;
+		position->angleOffsetFromTriggerEvent = offsetFromTriggerEvent;
 	}
 }
 
@@ -547,6 +550,11 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e ambiguousOperat
 
 	case TT_VVT_NISSAN_VQ35:
 		initializeNissanVQvvt(this);
+		break;
+
+	case TT_NISSAN_QR25:
+	case TT_TEMP_62:
+		initializeNissanQR25crank(this);
 		break;
 
 	case TT_NISSAN_VQ30: