+++ firmware/hw_layer/adc/adc_inputs.cpp
@@ -19,24 +19,15 @@
  * @author Andrey Belomutskiy, (c) 2012-2020
  */
 
-#include "global.h"
+#include "pch.h"
 
 #if HAL_USE_ADC
 #include "os_access.h"
 
-#include "engine.h"
-#include "adc_inputs.h"
 #include "adc_subscription.h"
 #include "AdcConfiguration.h"
 #include "mpu_util.h"
 #include "periodic_thread_controller.h"
-
-#include "pin_repository.h"
-#include "engine_math.h"
-#include "engine_controller.h"
-#include "maf.h"
-#include "perf_trace.h"
-#include "thread_priority.h"
 
 /* Depth of the conversion buffer, channels are sampled X times each.*/
 #ifndef ADC_BUF_DEPTH_FAST
@@ -114,7 +105,12 @@ static adcsample_t getAvgAdcValue(int index, adcsample_t *samples, int bufDepth,
 #define ADC_SAMPLING_FAST ADC_SAMPLE_28
 
 #if EFI_USE_FAST_ADC
-void adc_callback_fast(ADCDriver *adcp);
+static void adc_callback_fast(ADCDriver *adcp) {
+	// State may not be complete if we get a callback for "half done"
+	if (adcp->state == ADC_COMPLETE) {
+		onFastAdcComplete(adcp->samples);
+	}
+}
 
 static ADCConversionGroup adcgrpcfgFast = {
 	.circular			= FALSE,