+++ firmware/hw_layer/mmc_card.cpp
@@ -12,7 +12,7 @@
  * todo: extract some logic into a controller file
  */
 
-#include "global.h"
+#include "pch.h"
 
 #if EFI_FILE_LOGGING
 
@@ -20,14 +20,11 @@
 #include <stdio.h>
 #include <string.h>
 #include "mmc_card.h"
-#include "pin_repository.h"
 #include "ff.h"
 #include "hardware.h"
-#include "engine_configuration.h"
 #include "status_loop.h"
 #include "buffered_writer.h"
 #include "mass_storage_init.h"
-#include "thread_priority.h"
 
 #include "rtc_helper.h"
 
@@ -394,6 +391,11 @@ static BaseBlockDevice* initializeMmcBlockDevice() {
 static bool mountMmc() {
 	auto cardBlockDevice = initializeMmcBlockDevice();
 
+#if EFI_TUNER_STUDIO
+	// If not null, card is present
+	tsOutputChannels.sd_present = cardBlockDevice != nullptr;
+#endif
+
 #if HAL_USE_USB_MSD
 	// Wait for the USB stack to wake up, or a 5 second timeout, whichever occurs first
 	msg_t usbResult = usbConnectedSemaphore.wait(TIME_MS2I(5000));
@@ -479,6 +481,10 @@ static THD_FUNCTION(MMCmonThread, arg) {
 		// no card present (or mounted via USB), don't do internal logging
 		return;
 	}
+
+	#if EFI_TUNER_STUDIO
+		tsOutputChannels.sd_logging_internal = true;
+	#endif
 
 	while (true) {
 		// if the SPI device got un-picked somehow, cancel SD card