+++ firmware/hw_layer/sensors/cj125.cpp
@@ -14,15 +14,11 @@
  *
  */
 
-#include "engine.h"
+#include "pch.h"
+
 #include "cj125.h"
-#include "pwm_generator_logic.h"
-#include "rpm_calculator.h"
-#include "thread_priority.h"
 
 #if EFI_CJ125
-
-#include "adc_inputs.h"
 
 #if HAL_USE_SPI
 #include "mpu_util.h"
@@ -33,7 +29,6 @@
 
 #include "hardware.h"
 #include "backup_ram.h"
-#include "pin_repository.h"
 
 static CJ125 globalInstance;