+++ firmware/init/sensor/init_tps.cpp
@@ -1,8 +1,6 @@
-#include "global.h"
-#include "adc_inputs.h"
+#include "pch.h"
+
 #include "adc_subscription.h"
-#include "engine.h"
-#include "error_handling.h"
 #include "functional_sensor.h"
 #include "redundant_sensor.h"
 #include "redundant_ford_tps.h"