+++ firmware/CHANGELOG.md
@@ -27,6 +27,15 @@ All notable user-facing or behavior-altering changes will be documented in this 
 
 ## Month 202x Release - "Release Name"
 
+### Added
+ - RUSEFI MSD now contains bundle-specific URL #2848
+ - Nissan 4 cylinder QR trigger wheel #3118
+
+## Removed
+ -  TS file read feature #3004
+
+## July 2021 Release "Ice Cream Day"
+
 ### Breaking Changes
  - vvtOffset field migrated to four vvtOffsets fields. Anyone using VVT would need to manually adjust their configuration.  
 
@@ -37,7 +46,8 @@ All notable user-facing or behavior-altering changes will be documented in this 
  - new hardware testing console commands #2978
  
 ### Fixed
- - rusEFI console start-up speed #2964   
+ - rusEFI console start-up speed #2964
+ - wrong handling of symmetrical crank wheels NB2 VQ35 Renix #2980   
 
 ## June 2021 Release "National Logistics Day"